refactor(forge): split GitHub merge-request methods into github_mr.go (#1025)
Pure same-package code motion: move GetMergeRequest, toGitHubMergeRequest,
githubMRState, ListMergeRequestComments, reviewThreadIDsByDatabaseID,
reviewThreadCommentDBIDs, ReplyMergeRequestComment, and
ResolveMergeRequestThread verbatim out of github.go into the new github_mr.go.
graphqlDo stays in github_graphql.go (from M1); github_mr.go calls it
cross-file. No behaviour, signature, body, or doc-comment change.

Dropped the stale free-standing section header
"// --- Merge requests, issue events, notes, token introspection (M3) ---------"
rather than moving it: it named four concerns now split across four seam
files (issue events -> github_labels.go, notes -> github_issues.go, token
introspection -> github_auth.go in M1, merge requests -> github_mr.go), so it
documents no single block and would be orphaned over no code in github.go.
The mandated per-file seam header comment replaces it. It is a free-standing
comment, so go doc -u -all is unchanged (byte-identical to the pre-M1
baseline).
